Fix anopa build in morty.
The override of CFLAGS was causing us to lose the HASH directives, which
subsequently caused us to fail QA testing.

Since we don't need to debug any more, and morty builds include the '-g' anyway,
we can just remove that. We do have to force the exec-prefix to / now to stay
compatible with our previous builds.