s6rc-baselaytout: don't fail to mount proc if already mounted.
When booting with a stage0 (e.g. an initramfs loading us on disk), filesystems
like /proc tend to already be mounted. When this happens, attempting to mount
again causes failure, which can result in s6-rc not bringing up the rest of
the system.

Now we only mount it if there is no /proc/self, which seems like a decent enough test.