WIP bootable core-image-s6rc-small
This "Work in Progress" produces an image that runs under sumo with:
 bitbake core-image-s6rc-small
 runqemu qemux86 core-image-s6rc-small kvm

However, s6-linux-init-maker is not designed to run on any machine other
than the target, so it's not suitable for building the layout in yocto
(or at least I haven't successfully gotten it to). Even if I did, it
wouldn't be so good if yocto is cross-compiling. I should be able to
reproduce what it does but in the baselayout recipe, using files from
the target s6-linux-init package (that is, not using the -native one).
That work hasn't been done yet.

As such, after this image boots, you'll find youself in a shell with no
job control, where you'll need to follow the quick-start guide for
s6-linux-init-maker and run it yourself to make the system fully usable,
as well I think changing one or two of the services to not dump errors
to the console every second (can't remember which one at the moment, but
I hope to have that fixed in the next round.

There's also some extra setup that's needed in order for e.g. dropbear
to work.

So, TODO for this WIP:
1: Emulate s6-linux-init-maker to get the base set of files it makes
   into the final image
2: fix /var/log to be a directory
3: add empty /var/log/lastlog (for login message) and /var/log/wtmp (for
   'last' command)
4: add service to create /run/utmp on boot (for wall, who, w commands)
5: add service to create /dev/pts and mount a devpts filesystem on it
   (for dropbear to create pseudo terminals for new connections)
6: Fix error messages for whatever that service was that was spamming
   the console
7: Add a dropbear service example. I was running `dropbear -RFE` in my
   test environment. Should probably get its output logged somewhere