Booting s6rc-small
Did the linux-maker parts 'by hand'. Updated README to reflect current
state of things.

TODO: LiveCD doesn't work because initramfs is mounting /sys and /proc
already, but s6rc units also want to do that but fail because it's
"busy". I _think_ we're still trying to run the anopa initramfs; should
play with that a bit to see if we can make a general one or an s6rc
specific one.

TODO: Still want to start dropbear for sample image.